+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Made default game folder name depend on game edition
+- Improved game repairing feature
+- Replaced `curl` dependency by native code
+
 ## [3.3.0] - 24.03.2023
 
 ### Added